refactor: rename hintBaseSchema to hintParamsSchema
