+++ src/hints/hintParamsSchema.ts
