+++ src/hints/testLessThan.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 import { Felt } from 'primitives/felt';
 import { VirtualMachine } from 'vm/virtualMachine';
-import { cellRef, resOp, CellRef, ResOp } from 'hints/hintBaseSchema';
+import { cellRef, resOp, CellRef, ResOp } from 'hints/hintParamsSchema';
 import { HintName } from 'hints/hintName';
 
 /** Zod object to parse TestLessThan hint */
